+++ extension.js
@@ -41,6 +41,7 @@ function getHtml(webview, nonce, context, document, docBaseUri, settings) {
 <body class="mode-${initialMode}">
     <div id="app"></div>
     <script nonce="${nonce}" src="${mediaUri('marked.min.js')}"></script>
+    <script nonce="${nonce}" src="${mediaUri('turndown.min.js')}"></script>
     <script nonce="${nonce}" src="${mediaUri('katex.min.js')}"></script>
     <script nonce="${nonce}" src="${mediaUri('katex-auto-render.min.js')}"></script>
     <script nonce="${nonce}" src="${mediaUri('highlight.min.js')}"></script>